+++ src/main/java/baseball/service/ui/Input.java
@@ -1,4 +1,4 @@
-package baseball.service;
+package baseball.service.ui;
 
 import baseball.domain.Game;
 import utils.InputConverter;
