+++ src/main/java/baseball/service/ui/Output.java
@@ -1,4 +1,4 @@
-package baseball.service;
+package baseball.service.ui;
 
 import baseball.domain.Number;
